vss lab pc/user-authentication-service/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
pyjwt==2.8.0
werkzeug==2.3.7
bcrypt==4.1.1
flask-cors==4.0.0


